Enemy States Fixed

+++ Assets/Scripts/EnemyTank/EnemyScriptableObject.cs
@@ -8,7 +8,7 @@ public class EnemyScriptableObject : ScriptableObject
     public float range;
     public ParticleSystem deathExplosion;
     public AudioClip shootClip;
-    public int sightRange;
-    public int attackRange;
+    public float sightRange;
+    public float attackRange;
 
 }
